Changes for PR review: 1. resolve merge conflicts 2. rename struct `CFG` to `LIR` 3. rename directory from `ssa_ir` to `lir` 3. remove commented code
