+++ src/lir/ssa_type.rs
@@ -4,7 +4,7 @@ use std::fmt;
 
 use crate::sema::ast;
 use crate::sema::ast::ArrayLength;
-use crate::ssa_ir::expressions::Operand;
+use crate::lir::expressions::Operand;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum StructType {
