--- src/ssa_ir/mod.rs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod cfg;
-pub mod converter;
-pub mod expressions;
-pub mod instructions;
-pub mod printer;
-pub mod ssa_type;
-pub mod vartable;
