+++ tests/lir.rs
@@ -1,0 +1,1 @@
+mod lir_tests;
