+++ tests/lir_tests/insn_to_string.rs
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::ssa_ir_tests::helpers::{identifier, new_printer, new_vartable, num_literal};
+use crate::lir_tests::helpers::{identifier, new_printer, new_vartable, num_literal};
 use crate::{new_printer, num_literal, stringfy_insn};
 use num_bigint::BigInt;
 use solang::codegen::cfg;
 use solang::sema::ast::{ArrayLength, CallTy};
-use solang::ssa_ir::expressions::{BinaryOperator, Expression};
-use solang::ssa_ir::instructions::Instruction;
-use solang::ssa_ir::ssa_type::{InternalCallTy, PhiInput, StructType, Type};
+use solang::lir::expressions::{BinaryOperator, Expression};
+use solang::lir::instructions::Instruction;
+use solang::lir::ssa_type::{InternalCallTy, PhiInput, StructType, Type};
 use solang_parser::pt::Loc;
 
 #[test]
@@ -62,7 +62,7 @@ fn test_stringfy_returncode_insn() {
 #[test]
 fn test_stringfy_set_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(121, &Type::Uint(8));
     v.set_tmp(122, &Type::Uint(8));
     let printer = new_printer(v);
@@ -88,7 +88,7 @@ fn test_stringfy_set_insn() {
 #[test]
 fn test_stringfy_store_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(0, &Type::Ptr(Box::new(Type::Uint(8))));
     v.set_tmp(1, &Type::Uint(8));
     let printer = new_printer(v);
@@ -120,7 +120,7 @@ fn test_stringfy_store_insn() {
 #[test]
 fn test_stringfy_push_memory_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(
         3,
         &Type::Ptr(Box::new(Type::Array(
@@ -146,7 +146,7 @@ fn test_stringfy_push_memory_insn() {
 #[test]
 fn test_stringfy_pop_memory_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(
         3,
         &Type::Ptr(Box::new(Type::Array(
@@ -174,11 +174,10 @@ fn test_stringfy_pop_memory_insn() {
 #[test]
 fn test_stringfy_load_storage_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(101, &Type::Uint(32));
     v.set_tmp(3, &Type::StoragePtr(false, Box::new(Type::Uint(32))));
     let printer = new_printer(v);
-    // "%{} = load_storage slot({}) ty:{};"
     assert_eq!(
         stringfy_insn!(
             &printer,
@@ -194,7 +193,7 @@ fn test_stringfy_load_storage_insn() {
 #[test]
 fn test_stringfy_clear_storage_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(3, &Type::StoragePtr(false, Box::new(Type::Uint(32))));
     let printer = new_printer(v);
     assert_eq!(
@@ -211,7 +210,7 @@ fn test_stringfy_clear_storage_insn() {
 #[test]
 fn test_stringfy_set_storage_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::StoragePtr(false, Box::new(Type::Uint(256))));
     let printer = new_printer(v);
     assert_eq!(
@@ -229,11 +228,10 @@ fn test_stringfy_set_storage_insn() {
 #[test]
 fn test_stringfy_set_storage_bytes_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Bytes(32));
     v.set_tmp(2, &Type::StoragePtr(false, Box::new(Type::Bytes(32))));
     let printer = new_printer(v);
-    // set_storage_bytes {} offset:{} value:{}
     assert_eq!(
         stringfy_insn!(
             &printer,
@@ -250,7 +248,7 @@ fn test_stringfy_set_storage_bytes_insn() {
 #[test]
 fn test_stringfy_push_storage_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(101, &Type::Uint(32));
     v.set_tmp(
         3,
@@ -272,7 +270,6 @@ fn test_stringfy_push_storage_insn() {
                 storage: identifier(3)
             }
         ),
-        // "%101 = push_storage %3 uint32(1);"
         "uint32 %temp.ssa_ir.101 = push_storage storage_ptr<uint32[3]>(%temp.ssa_ir.3) uint32(1);"
     );
 }
@@ -280,7 +277,7 @@ fn test_stringfy_push_storage_insn() {
 #[test]
 fn test_stringfy_pop_storage_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(123, &Type::Uint(32));
     v.set_tmp(
         3,
@@ -301,7 +298,6 @@ fn test_stringfy_pop_storage_insn() {
                 storage: identifier(3)
             }
         ),
-        // "%123 = pop_storage %3;"
         "uint32 %temp.ssa_ir.123 = pop_storage storage_ptr<uint32[3]>(%temp.ssa_ir.3);"
     );
 
@@ -313,7 +309,6 @@ fn test_stringfy_pop_storage_insn() {
                 storage: identifier(3)
             }
         ),
-        // "pop_storage %3;"
         "pop_storage storage_ptr<uint32[3]>(%temp.ssa_ir.3);"
     )
 }
@@ -321,7 +316,7 @@ fn test_stringfy_pop_storage_insn() {
 #[test]
 fn test_stringfy_call_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Uint(8));
     v.set_tmp(2, &Type::Uint(64));
     v.set_tmp(3, &Type::Uint(8));
@@ -343,7 +338,6 @@ fn test_stringfy_call_insn() {
                 args: vec![num_literal!(3), identifier(133), num_literal!(6, 64)],
             }
         ),
-        // "%1, %2, %3 = call builtin#123(uint8(3), %133, uint64(6));"
         "uint8 %temp.ssa_ir.1, uint64 %temp.ssa_ir.2, uint8 %temp.ssa_ir.3 = call builtin#123(uint8(3), uint64(%temp.ssa_ir.133), uint64(6));"
     );
 
@@ -356,7 +350,6 @@ fn test_stringfy_call_insn() {
                 args: vec![num_literal!(3), identifier(133), num_literal!(6, 64)],
             }
         ),
-        // "%1, %2, %3 = call %123(uint8(3), %133, uint64(6));"
         "uint8 %temp.ssa_ir.1, uint64 %temp.ssa_ir.2, uint8 %temp.ssa_ir.3 = call ptr<fn(uint8, uint64, uint64) -> (uint8, uint64, uint8)>(%temp.ssa_ir.123)(uint8(3), uint64(%temp.ssa_ir.133), uint64(6));"
     );
 
@@ -369,7 +362,6 @@ fn test_stringfy_call_insn() {
                 args: vec![num_literal!(3), identifier(133), num_literal!(6, 64)],
             }
         ),
-        // "%1, %2, %3 = call function#123(uint8(3), %133, uint64(6));"
         "uint8 %temp.ssa_ir.1, uint64 %temp.ssa_ir.2, uint8 %temp.ssa_ir.3 = call function#123(uint8(3), uint64(%temp.ssa_ir.133), uint64(6));"
     );
 }
@@ -377,23 +369,8 @@ fn test_stringfy_call_insn() {
 //  ExternalCall
 #[test]
 fn test_stringfy_external_call_insn() {
-    // Insn::ExternalCall {
-    //     success,
-    //     address,
-    //     payload,
-    //     value,
-    //     accounts,
-    //     seeds,
-    //     gas,
-    //     callty,
-    //     contract_function_no,
-    //     flags,
-    // }
-
-    // {} = call_ext ty:{} address:{} payload:{} value:{} gas:{} accounts:{} seeds:{} contract_no:{}, function_no:{} flags:{};
-
     let mut v = new_vartable();
-    
+
     // success
     v.set_tmp(1, &Type::Bool);
     // payload
@@ -427,7 +404,7 @@ fn test_stringfy_external_call_insn() {
 #[test]
 fn test_stringfy_print_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(3, &Type::Uint(8));
     let printer = new_printer(v);
     assert_eq!(
@@ -444,7 +421,7 @@ fn test_stringfy_print_insn() {
 #[test]
 fn test_stringfy_memcopy_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(3, &Type::Bytes(32));
     v.set_tmp(4, &Type::Bytes(16));
     let printer = new_printer(v);
@@ -457,7 +434,6 @@ fn test_stringfy_memcopy_insn() {
                 bytes: num_literal!(16)
             }
         ),
-        // "memcopy %3 to %4 for uint8(16) bytes;"
         "memcopy bytes32(%temp.ssa_ir.3) to bytes16(%temp.ssa_ir.4) for uint8(16) bytes;"
     )
 }
@@ -465,7 +441,7 @@ fn test_stringfy_memcopy_insn() {
 #[test]
 fn test_stringfy_value_transfer_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Bool);
     v.set_tmp(
         2,
@@ -492,7 +468,7 @@ fn test_stringfy_value_transfer_insn() {
 #[test]
 fn test_stringfy_selfdestruct_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(
         3,
         &Type::Ptr(Box::new(Type::Struct(StructType::UserDefined(0)))),
@@ -512,12 +488,11 @@ fn test_stringfy_selfdestruct_insn() {
 #[test]
 fn test_stringfy_emit_event_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Bytes(32));
     v.set_tmp(2, &Type::Bytes(32));
     v.set_tmp(3, &Type::Bytes(32));
     let printer = new_printer(v);
-    // emit event#{} to topics[{}], data: {};
     assert_eq!(
         stringfy_insn!(
             &printer,
@@ -542,10 +517,9 @@ fn test_stringfy_branch_insn() {
 #[test]
 fn test_stringfy_branch_cond_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(3, &Type::Bool);
     let printer = new_printer(v);
-    // cbr {} block#{} else block#{};
     assert_eq!(
         stringfy_insn!(
             &printer,
@@ -555,7 +529,6 @@ fn test_stringfy_branch_cond_insn() {
                 false_block: 6
             }
         ),
-        // "cbr %3 block#5 else block#6;"
         "cbr bool(%temp.ssa_ir.3) block#5 else block#6;"
     )
 }
@@ -563,7 +536,7 @@ fn test_stringfy_branch_cond_insn() {
 #[test]
 fn test_stringfy_switch_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Uint(8));
     v.set_tmp(4, &Type::Uint(8));
     v.set_tmp(5, &Type::Uint(8));
@@ -581,10 +554,8 @@ fn test_stringfy_switch_insn() {
             default: 14,
         }
     );
-    // println!("{}", s);
     assert_eq!(
         s,
-        // "switch %1 cases: [%4 => block#11, %5 => block#12, %6 => block#13] default: block#14;"
         r#"switch uint8(%temp.ssa_ir.1):
     case:    uint8(%temp.ssa_ir.4) => block#11, 
     case:    uint8(%temp.ssa_ir.5) => block#12, 
@@ -596,7 +567,7 @@ fn test_stringfy_switch_insn() {
 #[test]
 fn test_stringfy_return_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Uint(8));
     v.set_tmp(2, &Type::Bytes(32));
     let printer = new_printer(v);
@@ -614,7 +585,7 @@ fn test_stringfy_return_insn() {
 #[test]
 fn test_stringfy_assert_failure_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(3, &Type::Bytes(32));
     let printer = new_printer(v);
     assert_eq!(
@@ -658,7 +629,7 @@ fn test_stringfy_unimplemented_insn() {
 #[test]
 fn test_stringfy_phi_insn() {
     let mut v = new_vartable();
-    
+
     v.set_tmp(1, &Type::Uint(8));
     v.set_tmp(2, &Type::Uint(8));
     v.set_tmp(12, &Type::Uint(8));
@@ -674,7 +645,6 @@ fn test_stringfy_phi_insn() {
                 ],
             }
         ),
-        // "%12 = phi [%1, block#13], [%2, block#14];"
         "uint8 %temp.ssa_ir.12 = phi [uint8(%temp.ssa_ir.1), block#13], [uint8(%temp.ssa_ir.2), block#14];"
     )
 }
