+++ tests/lir_tests/mod.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
-mod cfg_to_string;
-mod convert_cfg;
+mod lir_to_string;
+mod convert_lir;
 mod expr_to_string;
 mod helpers;
 mod insn_to_string;
