--- tests/ssa_ir.rs
@@ -1,1 +1,0 @@
-mod ssa_ir_tests;
